JS: initiate.js 분리
